Environment: Split classes
Move Environment_Physical and Environment_Simulator into separate files
and make sure to only import them when needed. This is because the
physical environment now imports Distance_Sensor_Physical, and its
RPi.GPIO module can only used on the Pi, while the simulator uses VRML
which is not yet installed in the travis setup.

Fix all imports, with some simplification for Environment module import.
