+++ distance_sensor_physical.py
@@ -4,18 +4,19 @@ from __init__ import __package__
 from settings import Arguments
 from distance.Distance_Sensor_Physical import Distance_Sensor_Physical
 from geometry.Geometry import Geometry
-from trajectory.Environment import Environment_Simulator
+from environment import Environment
 from trajectory.MockVehicle import MockVehicle
 
 def main(argv):
     arguments = Arguments("settings.json", argv)
 
-    environment = Environment.setup(arguments)
-    distance_sensor = Distance_Sensor_Physical(environment)
+    environment = Environment.setup(arguments, simulated=False)
+    distance_sensors = environment.get_distance_sensors()
     settings = arguments.get_settings("distance_sensor_physical")
 
     while True:
-        print("Measured distance: {} m".format(distance_sensor.get_distance()))
+        for sensor in distance_sensors:
+            print("Measured distance: {} m".format(sensor.get_distance()))
         time.sleep(settings.get("interval_delay"))
 
 if __name__ == "__main__":
