+++ environment/Environment.py
@@ -1,7 +1,5 @@
-from VRMLLoader import VRMLLoader
-from ..distance.Distance_Sensor_Simulator import Distance_Sensor_Simulator
 from ..geometry import Geometry
-from MockVehicle import MockVehicle
+from ..trajectory.MockVehicle import MockVehicle
 
 class Environment(object):
     """
@@ -11,21 +9,23 @@ class Environment(object):
     _sensor_class = None
 
     @classmethod
-    def setup(self, arguments, geometry_class="Geometry", vehicle=None):
+    def setup(self, arguments, geometry_class="Geometry", vehicle=None, simulated=True):
         """
         Create an Environment object or simulated environment.
 
-        The returned object is an Enviromnent object or a subclass, loaded with the given `arguments` object. Optionally one can specify which `geometry_class` to use and what `vehicle` object to use.
-        Based on whether the `vehicle` is a `MockVehicle` (default), the environment is simulated or not.
+        The returned object is an Enviromnent object or a subclass, loaded with the given `arguments` object. Optionally one can specify which `geometry_class` to use and what `vehicle` object to use. To use an environment with physical distance sensors, set `simulated` to `False`.
+        By default, the `vehicle` is a `MockVehicle`.
         For more control over simulated environment setup, use the normal constructors instead.
         """
         geometry = Geometry.__dict__[geometry_class]()
         if vehicle is None:
             vehicle = MockVehicle(geometry)
 
-        if isinstance(vehicle, MockVehicle):
+        if simulated:
+            from Environment_Simulator import Environment_Simulator
             return Environment_Simulator(vehicle, geometry, arguments)
 
+        from Environment_Physical import Environment_Physical
         return Environment_Physical(vehicle, geometry, arguments)
 
     def __init__(self, vehicle, geometry, arguments):
@@ -94,53 +94,3 @@ class Environment(object):
         Get the pitch bearing of the vehicle.
         """
         return self.vehicle.attitude.pitch
-
-class Environment_Physical(Environment):
-    _sensor_class = Distance_Sensor_Physical
-
-class Environment_Simulator(Environment):
-    """
-    Simulated environment including objects around the vehicle and potentially the vehicle itself.
-    This allows us to simulate a mission without many dependencies on ArduPilot.
-    """
-
-    _sensor_class = Distance_Sensor_Simulator
-
-    def __init__(self, vehicle, geometry, arguments):
-        super(Environment_Simulator, self).__init__(vehicle, geometry, arguments)
-        scenefile = self.settings.get("scenefile")
-        translation = self.settings.get("translation")
-        if scenefile is not None:
-            loader = VRMLLoader(self, scenefile, translation)
-            self.objects = loader.get_objects()
-            return
-
-        # Use hardcoded objects for testing
-        l1 = self.get_location(100, 0, 10)
-        l2 = self.get_location(0, 100, 10)
-        l3 = self.get_location(-100, 0, 10)
-        l4 = self.get_location(0, -100, 10)
-
-        # Simplify function call
-        get_location_meters = self.geometry.get_location_meters
-        self.objects = [
-            {
-                'center': get_location_meters(self.vehicle.location, 40, -10),
-                'radius': 2.5,
-            },
-            (get_location_meters(l1, 40, -40), get_location_meters(l1, 40, 40),
-             get_location_meters(l1, -40, 40), get_location_meters(l1, -40, -40)
-            ),
-            (get_location_meters(l2, 40, -40), get_location_meters(l2, 40, 40),
-             get_location_meters(l2, -40, 40), get_location_meters(l2, -40, -40)
-            ),
-            (get_location_meters(l3, 40, -40), get_location_meters(l3, 40, 40),
-             get_location_meters(l3, -40, 40), get_location_meters(l3, -40, -40)
-            ),
-            (get_location_meters(l4, 40, -40), get_location_meters(l4, 40, 40),
-             get_location_meters(l4, -40, 40), get_location_meters(l4, -40, -40)
-            )
-        ]
-
-    def get_objects(self):
-        return self.objects
