+++ tests/distance_sensor_physical.py
@@ -1,7 +1,6 @@
 import unittest
 from mock import patch, call, MagicMock
 from ..settings import Arguments
-from ..trajectory.Environment import Environment
 
 class TestDistanceSensorPhysical(unittest.TestCase):
     def setUp(self):
@@ -15,11 +14,12 @@ class TestDistanceSensorPhysical(unittest.TestCase):
 
         self.patcher = patch.dict('sys.modules', modules)
         self.patcher.start()
+        from ..environment import Environment
         from ..distance.Distance_Sensor_Physical import Distance_Sensor_Physical
-        arguments = Arguments("settings.json", [])
+        arguments = Arguments("settings.json", ["--sensors", "0"])
         self.settings = arguments.get_settings("distance_sensor_physical")
-        environment = Environment.setup(arguments)
-        self.distance_sensor = Distance_Sensor_Physical(environment)
+        environment = Environment.setup(arguments, simulated=False)
+        self.distance_sensor = environment.get_distance_sensors()[0]
 
     def tearDown(self):
         self.patcher.stop()
