+++ trajectory/Monitor.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 from matplotlib.patches import Polygon, Circle
 from matplotlib.collections import PatchCollection
 
-from Environment import Environment_Simulator
+from ..environment.Environment_Simulator import Environment_Simulator
 
 class Plot(object):
     """
